requirements.txt
asgiref==3.3.1
blis==0.7.4
catalogue==1.0.0
certifi==2020.12.5
cffi==1.14.4
chardet==4.0.0
cryptography==3.3.1
cycler==0.10.0
cymem==2.0.5
defusedxml==0.6.0
Django==3.1.4
django-allauth==0.44.0
djangorestframework==3.12.2
fastai==2.1.10
fastcore==1.3.13
fastprogress==1.0.0
idna==2.10
joblib==1.0.0
kiwisolver==1.3.1
matplotlib==3.3.3
murmurhash==1.0.5
numpy==1.19.4
oauthlib==3.1.0
packaging==20.8
pandas==1.2.0
Pillow==8.0.1
plac==1.1.3
preshed==3.0.5
pycparser==2.20
PyJWT==2.0.0
pyparsing==2.4.7
python-dateutil==2.8.1
python-dotenv==0.15.0
python3-openid==3.2.0
pytz==2020.5
PyYAML==5.3.1
requests==2.25.1
requests-oauthlib==1.3.0
scikit-learn==0.24.0
scipy==1.6.0
six==1.15.0
spacy==2.3.5
sqlparse==0.4.1
srsly==1.0.5
thinc==7.4.5
threadpoolctl==2.1.0
torch==1.7.1+cpu
torchaudio==0.7.2
torchvision==0.8.2+cpu
tqdm==4.55.0
typing-extensions==3.7.4.3
urllib3==1.26.2
wasabi==0.8.0
